requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
Markdown==3.9
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
django-silk==5.4.3
requests==2.32.5
PyJWT==2.10.1
djangorestframework_simplejwt==5.5.1
django_celery_results==2.6.0
django-redis==6.0.0
celery==5.5.3
django-celery-beat==2.8.1
python-crontab==3.3.0
django-ckeditor==6.7.3
pillow==12.0.0
phonenumbers==9.0.18
sslcommerz-lib==1.0